requirements.txt
pygame>=2.5.0
Pillow>=10.0.0
customtkinter>=5.2.0
numpy>=1.24.0
